add (HACKY) support for params in memory
Hacky because we dont support the evaluating the frame base expression
yet. We have to decide whether we want to store the expression in the
parameter list of function. If we decide that, we have to switch to a
arena based approach for gimli. Do we want to keep the entire binaries
around for the lifetime of the program?